+++ src/Taxes/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 6.x Series
 
-## Unreleased
-##### 2026-XX-YY
+## 6.0.0
+##### 2026-09-25
 
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support